fix: hide mfa types
