+++ src/GoTrueClient.ts
@@ -88,12 +88,13 @@ import type {
   LockFunc,
   UserIdentity,
   SignInAnonymouslyCredentials,
-  MFAEnrollTOTPParams,
-  AuthMFAEnrollTOTPResponse,
-  AuthMFAEnrollErrorResponse,
-  MFAEnrollPhoneParams,
-  AuthMFAEnrollPhoneResponse,
 } from './lib/types'
+import {
+  MFAEnrollTOTPParams,
+  MFAEnrollPhoneParams,
+  AuthMFAEnrollTOTPResponse,
+  AuthMFAEnrollPhoneResponse,
+} from './lib/internal-types'
 
 polyfillGlobalThis() // Make "globalThis" available
 
@@ -313,9 +314,7 @@ export default class GoTrueClient {
         if (error) {
           this._debug('#_initialize()', 'error detecting session from URL', error)
 
-          if (
-            error?.code === 'identity_already_exists'
-          ) {
+          if (error?.code === 'identity_already_exists') {
             return { error }
           }
 
@@ -2355,12 +2354,8 @@ export default class GoTrueClient {
   /**
    * {@see GoTrueMFAApi#enroll}
    */
-  private async _enroll(
-    params: MFAEnrollTOTPParams
-  ): Promise<AuthMFAEnrollTOTPResponse | AuthMFAEnrollErrorResponse>
-  private async _enroll(
-    params: MFAEnrollPhoneParams
-  ): Promise<AuthMFAEnrollPhoneResponse | AuthMFAEnrollErrorResponse>
+  private async _enroll(params: MFAEnrollTOTPParams): Promise<AuthMFAEnrollTOTPResponse>
+  private async _enroll(params: MFAEnrollPhoneParams): Promise<AuthMFAEnrollPhoneResponse>
   private async _enroll(params: MFAEnrollParams): Promise<AuthMFAEnrollResponse> {
     try {
       return await this._useSession(async (result) => {
